Updates to PN532.cpp
Changes lines to correct warnings of unused variables (now returned in their appropriate functions) and corrected int types where needed.

+++ example_emulate_tag/example_emulate_tag.ino
@@ -18,7 +18,7 @@ EmulateTag nfc(pn532hsu);
 
 uint8_t ndefBuf[120];
 NdefMessage message;
-int messageSize;
+uint8_t messageSize;
 
 uint8_t uid[3] = {0x12, 0x34, 0x56};
